Remove logging to file and json
The web portal does not need any logfile to be created to work.

This allows the removal of the VOLUME, the need to have a dynamically
allocated user to run uwsgi under and being able to map the user
to a configurable one from outside the container.

Also because no logfile accumulate this removes the need to prune
them on a regular basis.